requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# Data Profiling
ydata-profiling==4.6.4
great-expectations==0.18.8

# Database Connectivity
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Visualization & Dashboard
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
requests==2.31.0
jinja2==3.1.2
tabulate==0.9.0

# Data Quality
dataclasses-json==0.6.3
pydantic==2.5.3
